Added missing files and a simple getting started manual for the demo using the infrastructure proposed in the REP-2017 below.
https://github.com/ros-infrastructure/rep/pull/385

The demo is to explain how to use the infrastructure in the process
that has multiple single-threaded executors as a reply to the question below.

https://github.com/ros-infrastructure/rep/pull/385#discussion_r1350512216

The YAML files for the existing executors are to demonstrate that
the existing executors can be enhanced with our proposed infrastructure,
supporting our mention below.

https://github.com/ros-infrastructure/rep/pull/385#discussion_r1347295686

Signed-off-by: Shoji Morita <s-morita@esol.co.jp>